updated code to work with new ui
