+++ client/src/components/search/SearchPage.js
@@ -4,15 +4,15 @@ import SearchResults from './SearchResults';
 
 export default function SearchPage(props) {
   const {
-    location: {
-      state: { results },
-    },
+    location: { state },
   } = props;
   return (
     <div>
       <h3>Stories or Tags or People</h3>
       <div>here is a complete list of search results</div>
-      <SearchResults results={results} limit={results.length} />
+      {!_.isEmpty(state) && (
+        <SearchResults results={state.results} limit={state.results.length} />
+      )}
     </div>
   );
 }
